default-text-722-1

+++ src/App.js
@@ -5,24 +5,47 @@ import ReactMarkdown from "react-markdown";
 
 function App() {
 
-  const [text, setText] = useState('');
+  const [text, setText] = useState(`
+## Markdown editor & previewer
+---
+- I have created this tool to edit README.md files for GitHub.
+- Node.js 18.16.1, ReactJS, react-markdown   
+---
+
+### How to use
+
+- Type the markdown text into the editor
+- Use the markdown syntax
+- Text will reload to the editor text area the next time you will take a visit
+- You can clear the text from the editor area using the clear button at the top
+- This tool is not suitable for smartphones or any other devices with small screens
+
+### Links
+
+[markdown editor](https://markdown-editor.up.railway.app/)
+
+[markdown syntax](https://www.markdownguide.org/basic-syntax/)`);
+  
+  const storedText = localStorage.getItem('text');
 
   useEffect(() => {
     // Load the text from localStorage
-    const storedText = localStorage.getItem('text');
     if (storedText) {
       setText(storedText);
     }
   }, []);
 
   const handleChange = (e) => {
+    // Save text (type) to the localStorage 
     setText(e.target.value);
     localStorage.setItem('text', e.target.value);
   }
 
   const clearButton = (e) => {
-    localStorage.clear('text');
-    window.location.reload();
+    if (storedText){
+      localStorage.clear('text');
+    }
+    setText('');
   }
 
   return (
